+++ notification/src/main/java/com/microservice/erp/services/iServices/ISendNotificationService.java
@@ -3,8 +3,10 @@ package com.microservice.erp.services.iServices;
 import com.microservice.erp.domain.dto.NoticeDto;
 import org.springframework.http.ResponseEntity;
 
+import java.math.BigInteger;
+
 public interface ISendNotificationService {
     ResponseEntity<?> sendNotification(String authHeader, NoticeDto noticeDto) throws Exception;
 
-    ResponseEntity<?> checkNoticeAlreadySentOrNot(String year, Long noticeConfigurationId);
+    ResponseEntity<?> checkNoticeAlreadySentOrNot(String year, BigInteger noticeConfigurationId);
 }
